+++ src/pages/home/Homepage.css
@@ -5,86 +5,4 @@
     max-width: 100vw;
     max-height: 100vh;
 }
-.top{
-    display: flex;
-    border-radius: 30px;
-    width: 1110px;
-    height: 197px;
-    justify-content: center;
-    align-items: center;
-    box-shadow: 0 0 20px 0 rgba(0, 0, 0, 0.1);
-    background: #101010;
-    margin-top: 22px;
-    box-sizing: border-box;
-    margin: 0px 165px 0px 165px;
-}
-.top__text{
-    color: #fff;
-    font-weight: 600;
-    font-size: 30px;
-}
-.top__text-title{
-    display: flex;
-    gap: 12px;
-    justify-content: center;
-}
-.top__text-h1{
-    font-weight: 600;
-    font-size: 30px;
-}
-.headphones{
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    gap: 30px;
-}
-.headphones__card{
-    display: flex;
-    flex-direction: column;
-    border-radius: 30px;
-    width: 350px;
-    height: 407px;
-    box-shadow: 0 0 20px 0 rgba(0, 0, 0, 0.1);
-    background: #fff;
-    padding: 15px 21px 32px 21px;
-}
-.headphones__card-top{
-    display: flex;
-    gap: 24px;
-}
-.headphones__favorite{
-    border: 0;
-    width: 20px;
-    height: 20px;
-    background-color: #fff;
-}
-.headphones__card-img{
-    width: 219px;
-    height: 237px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-.headphones__info{
-    display: flex;
-    justify-content: space-between;
-    margin-top: 55px;
-}
-.headphones__info-title{
-    font-weight: 600;
-    font-size: 17px;
-    color: #1c1c27;
-}
-.headphones__info-price{
-    font-weight: 600;
-    font-size: 17px;
-    color: #ffa542;
-}
-.headphones__estimation{
-    display: flex;
-    margin-top: 9px;
-}
-.headphones__estimation-text{
-    font-weight: 600;
-    font-size: 17px;
-    color: #838383;
-}
+
