+++ src/pages/home/interfaces.ts
@@ -1,0 +1,11 @@
+export interface Idata {
+    category: string;
+    description: string;
+    estimation: string;
+    img: string;
+    inBasket: boolean;
+    inFavorite: boolean;
+    name: string;
+    priceAfterDiscount: string;
+    priceBeforeDiscount: string;
+}
