+++ src/pages/home/sections/headphones/index.ts
@@ -1,0 +1,2 @@
+import Headphones from "./Headphones";
+export default Headphones
