+++ src/pages/home/sections/top/Top.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import './Top.css'
+
+const Top: React.FC = () => {
+    return (
+        <section>
+            <div className="top">
+                <div className="top__text">
+                    <div className="top__text-title">
+                        <h1  className="top__text-h1">Аксессуары</h1>
+                        <p  className="top__text-p">для </p>
+                    </div>
+                    <p className="top__text-sub">Iphone 13 Pro Max</p>
+                </div>
+                {/* <img src="../../../images/iPhone-13-Pro-Max-silver-1000x1000 1.jpg" alt="phone" /> */}
+            </div>`
+        </section>
+    )
+}
+export default Top
