+++ src/pages/home/sections/top/index.ts
@@ -1,0 +1,2 @@
+import Top from "./Top";
+export default Top
